+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version 3.4.0 *(In Development)*
  * Library now uses the `r4` version of the compatibility library for its base.
  * Context menu callbacks now use the support version of `MenuItem` to maintain
    consistency.
+ * Added preference plugin which provides an action bar enhanced preference
+   screen.
 
 
 Version 3.3.1 *(2011-10-20)*